Move file analysers
